put search and filters back in a single thing
