+++ react/javascript/src/components/app/FilteredResults.module.scss
@@ -2,7 +2,6 @@
     margin-bottom: 1rem;
 
     > * {
-        border-top-width: 0;
-        border-bottom-width: 0;
+        border-top: 0 !important;
     }
 }
