--- react/javascript/src/components/app/StatusesFilter.module.scss
@@ -1,21 +1,0 @@
-.filterLayout {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    padding: 1rem;
-    border: 1px solid rgba(0, 0, 0, 0.1);
-}
-
-.filterList {
-    display: flex;
-    align-items: center;
-    gap: 0.5rem;
-    list-style: none;
-    padding: 0;
-    margin: 0;
-}
-
-.filterField {
-    display: flex;
-    align-items: center;
-}
